backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.103.0
frozenlist==1.7.0
grpcio==1.74.0
h11==0.16.0
idna==3.10
multidict==6.6.4
packaging==25.0
propcache==0.3.2
protobuf==6.32.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.0
requests==2.32.5
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.23.2
xai-sdk==1.0.1
yarl==1.20.1
